fix: do not request partial encoded chunk from self
